Updates to Floyd's. Changed weight variable to a double in all graph subclasses and root class to be able to handle infinity value.

+++ ALWDG.java
@@ -20,7 +20,7 @@ public class ALWDG extends G{
 		G.outDegrees[i]++;
 	}
 	
-	protected void putEdge(int i, int j, int k) {
+	protected void putEdge(int i, int j, double k) {
 		Node node = new Node(j,k);
 		weightedList[i].add(node);
 		G.degrees[i]++;
